ShupStatus.csを作成
