--- SuperNewRoles/Patch/Harmony/ShipStatus.cs
@@ -1,13 +1,0 @@
-using HarmonyLib;
-
-namespace SuperNewRoles.Patch.Harmony
-{
-    [HarmonyPatch(typeof(ShipStatus), nameof(ShipStatus.Awake))]
-    class ShipStatusAwakePatch
-    {
-        private static void Postfix()
-        {
-            SuperNewRoles.Modules.ProctedMessager.Init();
-        }
-    }
-}
